+++ packages/react/src/themes/teams/components/Icon/svg/ProcessedIcons/icons-speaker-mute.tsx
