+++ packages/react/src/themes/teams/components/Icon/svg/ProcessedIcons/index.ts
@@ -96,7 +96,7 @@ import processedIcons_numberlist from './icons-number-list'
 import processedIcons_notificationoff from './icons-notification-off'
 import processedIcons_notes from './icons-notes'
 import processedIcons_nochat from './icons-no-chat'
-import processedIcons_noaudio from './icons-no-audio'
+import processedIcons_speakermute from './icons-speaker-mute'
 import processedIcons_newtab from './icons-newtab'
 import processedIcons_newcontactgroup from './icons-new-contactgroup'
 import processedIcons_myactivity from './icons-my-activity'
@@ -350,6 +350,7 @@ export default {
   processedIcons_star,
   processedIcons_speakerslashed,
   processedIcons_speakeroff,
+  processedIcons_speakermute,
   processedIcons_sketch,
   processedIcons_shareobject,
   processedIcons_share,
@@ -405,7 +406,6 @@ export default {
   processedIcons_notificationoff,
   processedIcons_notes,
   processedIcons_nochat,
-  processedIcons_noaudio,
   processedIcons_newtab,
   processedIcons_newcontactgroup,
   processedIcons_myactivity,
